This is the semi final version of homzy app made by team of Shashi Kant. It includes all the functionality a semi project has to be

+++ lib/screens/takedata_page.dart
@@ -218,7 +218,7 @@ print("asdfhlk ks");
                           print(widget.name);
                           print(widget.price);
 
-                          storeData(work:widget.name,price:widget.price);
+                          storeData(work:widget.name,price:widget.price.toInt());
                         },
                         child: Text('Submit'),
